Working towards MS2 ARM
Alright, this is finally much more in the right direction. Still have a fair amount to go though. Still need to make sure the message formats are what has been determined as the standard and debug the code. Good progress though. Let's see if I can make it happen before 0930.