ARM code update
Alright, here's another good update of the ARM code with most everything together for everyone to start working on things for their individual milestones. The README.md file includes a general TODO list I have for myself. Feel free to take a look at the code. I still have to run it through the IDE in the Win7 VM to work out errors and go through and visually see if I made any copy and paste errors, but I wanted to get a push up at this point before continuing on because I think everyone can get the general idea of who the code structure works from the code at this point if you all feel so inclined. Will keep at it and push updates as I can.