ARM code update - Nav Timer added
Added Seo's 10 ms timer for his Navigation task. Updated the README.md file's TODO list to reflect this.

This involved adding his timer to the myTimers file. Added a MsgType for the navigation task's switch statement. This was added in both the myTypes header file and the navigation.c file. The function for adding the timer interrupt to the navigation queue can be seen by the added function in the navigation.c file as well (I think that is the correct wording… I know it's what needs to be done for timers nonetheless because my motorControl task needs a timer as well since Seo and I don't get hardware driven timers per say like Richard and Dhiraj do).