ARM Code - Compiles & Almost Completely Framed Out
So, this commit has the ARM compiling on the VM with no errors
(obviously) and 2 warnings; the warnings are related to variables that
are going to be needed for when I finish working out the issues with
the ADC task and getting that working (the variables are related to the
LCD which we'll need to display, and in part prove, that we have I2C
and UART working between the ARM & 2 PICs. Will continue working on
getting the ADC things worked out. May not be unitl Sunday though.