ARM Code Major Update
This is a major update of the ARM code. This is untested in the VM IDE so I will be doing that after this commit / this evening sometime.

This adds all the necessary private functions for accessing data passed to everyones' tasks. The functions access specific array locations depending on which type of message is being processed by the task at that time. This is all based on the public functions everyones' tasks have & their respective message types. Looking at your task's header file and the myTypes.h file for a bit should make these make more sense for everyone. If there are still questions, feel free to ask me.